+++ CHANGELOG.md
@@ -10,11 +10,16 @@
   match. Previously only a class reference could be provided.
 * The Admin Console diff tool for Configs, Prefs, and JSONBlobs now displays who updated each value
   and when.
+* New "Hoist Inspector" tool available in Desktop apps for displaying and querying models running in
+  an application and their properties (especially observables).
+    * Powered by a new method `XH.activeModels()`, which supports listing and querying all models
+      instantiated within a running app, and `XH.inspectorService`, which provides additional
+      processing of that model data when active.
+    * Admin/dev-focused UI available on Desktop, activated via Admin-only options in the app menu
+      and version bar, or via `XH.inspectorService.activate()`.
 
 ### ⚙️ Technical
 
-* New method `XH.activeModels()` supports listing and querying all models instantiated within a
-  running app. Intended for use by devs looking to inspect or troubleshoot their models.
 * `HoistComponents` no longer mutate the props object passed to them in React production mode. This
   was not causing noticeable application issues, but could result in a component's base CSS class
   being applied multiple times to its DOM element.